AI bug fix
nextActionTime was preventing AI from marking current node set as unwalkable if they arrived when the timer was running causing a delay in grid update. If the AI is not moving we now update. Needs to be fixed to only update when stopped and again when moving rather than each frame.